+++ app/views/PlaybackAnswersView.scala.html
@@ -36,6 +36,15 @@
 
     @components.warning("playbackAnswers.declareAnnually")
 
+    @for(section <- trustDetails){
+        @{
+            section match {
+                case a: AnswerSection => components.summary_answer_section(a)
+                case r: RepeaterAnswerSection => components.repeater_answer_section(r)
+            }
+        }
+    }
+
     @for(section <- entities){
         @{
             section match {
@@ -47,14 +56,5 @@
 
     <p>@messages("playbackAnswers.informationFirstRegistered")</p>
     <p>@messages("playbackAnswers.selfAssessmentOnline")</p>
-
-    @for(section <- trustDetails){
-        @{
-            section match {
-                case a: AnswerSection => components.summary_answer_section(a)
-                case r: RepeaterAnswerSection => components.repeater_answer_section(r)
-            }
-        }
-    }
 
 }
