+++ app/views/VariationProgressView.scala.html
@@ -72,7 +72,7 @@
         <p>@messages("variationProgress.p3")</p>
         <h2 id="task-list--heading--summary" class="app-task-list__section">@messages("variationProgress.subHeading.4")</h2>
         <p>
-            @components.link(PlaybackDraftAnswersController.onPageLoad().url, "print-and-save", "variationProgress.printsave.link", openInNewWindow = true)
+            @components.link(controllers.print.routes.PrintMaintainDraftAnswersController.onPageLoad().url, "print-and-save", "variationProgress.printsave.link", openInNewWindow = true)
         </p>
     }-->
