+++ app/views/declaration/CloseTrustConfirmationView.scala.html
@@ -38,7 +38,7 @@
         </h1>
     </div>
 
-    <p class="panel-indent">@components.link(PlaybackDeclaredAnswersController.onPageLoad().url, "print-and-save","closeTrustConfirmation.printsave.link", openInNewWindow = true)</p>
+    <p class="panel-indent">@components.link(controllers.print.routes.PrintMaintainDeclaredAnswersController.onPageLoad().url, "print-and-save","closeTrustConfirmation.printsave.link", openInNewWindow = true)</p>
 
     <h2>@messages("closeTrustConfirmation.subheading")</h2>
