+++ app/views/declaration/ConfirmationView.scala.html
@@ -39,7 +39,7 @@
         </h1>
     </div>
 
-<p class="panel-indent">@components.link(PlaybackDeclaredAnswersController.onPageLoad().url, "print-and-save","confirmationPage.printsave.link", openInNewWindow = true)</p>
+<p class="panel-indent">@components.link(controllers.print.routes.PrintMaintainDeclaredAnswersController.onPageLoad().url, "print-and-save","confirmationPage.printsave.link", openInNewWindow = true)</p>
 
     <h2>@messages("confirmationPage.subheading1")</h2>
